% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_book_table.R
\name{as_field_book_table}
\alias{as_field_book_table}
\alias{read_fb}
\title{Field Book Tables}
\usage{
as_field_book_table(fbt, unique.id = "plot_id", row = "row",
  column = "column", line.name = "line_name", other.cols = NULL,
  check.names = TRUE)

read_fb(file, unique.id = "plot_id", row = "row", column = "column",
  line.name = "line_name", other.cols = NULL, check.names = TRUE)
}
\arguments{
\item{fbt}{A Field Book Table object or an object to convert into a Field
Book Table object.}

\item{unique.id}{The column name of the Field Book data frame that corresponds
to the unique identification of each entry in the data table. Defaults to 'plot_id.'}

\item{row}{The column name of the Field Book data frame that corresponds to
the row of each entry in the data table. Defaults to 'row.'}

\item{column}{The column name of the Field Book data frame that corresponds to
the column of each entry in the data table. Defaults to 'column.'}

\item{line.name}{The column name of the Field Book data frame that corresponds to
the column of each entry name in the data table. Defaults to 'line.name.'}

\item{other.cols}{A character vector of other non-trait columns.}

\item{check.names}{Logical whether the column names should be checked for
proper formatting.}

\item{file}{The path to the Field Book Table .csv to be read in.}
}
\value{
An object of class \code{data.frame} 
with formatting consistent with Field Book Table.
}
\description{
Creates/checks data.frame objects for Field Book Table format.

Reads in data from a Field Book Table as a \code{fbt} object.
}
\details{
\code{as_field_book_table} converts data to a \code{data.frame} object with
formatting consistent with the Field Book Table.

A Field Book Table is a .csv file obtained by exporting data from the Field
Book app (\url{http://wheatgenetics.org/field-book}). See the app 
manual for instructions for exporting data. To export a Field Book Table file, 
one must check the "Table" format and the "All Important Columns" boxes when 
exporting. Data from a Field Book Table file will have at least three required
columns: a unique identifier, the row, and the column. Remaining columns can
include other variables.
}
\section{Functions}{
\itemize{
\item \code{read_fb}: 
}}

