% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_heatmap.R
\name{fb_heatmap}
\alias{fb_heatmap}
\title{Generate a heatmap of field data}
\usage{
fb_heatmap(fbt, traits)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{\code{Character} column name of for which to plot heatmaps.
Defaults to all \code{numeric} traits.}
}
\description{
Creates a heatmap image for each numeric trait in the field_book_table 
file. An option for exporting a PDF of these images is provided. This function works best with
raw field_book_table files, so this could be used immediately after exporting
a file from the Field Book App.
}
