% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_outliers.R
\name{fb_outliers}
\alias{fb_outliers}
\title{Search for outliers in the field data}
\usage{
fb_outliers(fbt, traits, max.sd = 3)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A \code{character} of traits for which outliers should be 
discovered. By default all numeric traits are examined.}

\item{max.sd}{A \code{numeric} determining the number of standard deviations from
the mean to use when labelling outliers.}
}
\description{
Search for outliers in the field data
}
