% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_parse.R
\name{fb_parse}
\alias{fb_parse}
\title{Parse data from non-standard traits}
\usage{
fb_parse(fbt, traits, sep = ":", n.obs = 10)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A \code{character} vector of trait names identifying which column(s) in the
fbt object file to parse.}

\item{sep}{Character(s) used to split the multicat data entries.}

\item{n.obs}{An \code{numeric} vector of integers describing the number
of desired datapoint observations in a multicat trait. The first n observations will be
included and the remainder discarded For instance, if one of the
multicat data points was "1:2:3:4:5" and n.obs was set to 3, the resulting
data would be "1 2 3". The vector must be the same length as 'traits.'}
}
\value{
An fbt object with unparsed trait columns and with with individual columns for 
each datapoint entry in each multicat trait. The mean and standard deviation 
of all datapoint entries are also included.
}
\description{
Parse data from non-standard traits
}
\examples{
data("fbt_sample")

# Parse the 'Multi' trait
fbt1 <- fb_parse(fbt = fbt_sample, traits = "Multi", sep = ":")

}
