% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_traits.R
\name{fb_traits}
\alias{fb_traits}
\title{Extract the trait names from a Field Book Table}
\usage{
fb_traits(fbt, mode = c("all", "character", "numeric", "logical"))
}
\arguments{
\item{fbt}{A \code{field.book.table} object.}

\item{mode}{The class of traits to be returned. May be \code{"all"} (default),
\code{"character"}, \code{"numeric"} or \code{"logical"}.}
}
\description{
Returns a character vector of trait names of the desired mode
}
