% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_visualize.R
\name{fb_visualize}
\alias{fb_visualize}
\title{Visualize a summary of the field book data}
\usage{
fb_visualize(fbt, traits, graph = c("boxplot", "histogram"), bins = 30,
  nrow = NULL, ncol = NULL)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A character vector specifying the trait to plot. Defaults to all
numeric traits.}

\item{graph}{The type of plot to produce. Accepted values are \code{"boxplot"}
(default) or \code{"histogram"}.}

\item{bins}{The number of bins to force in the histogram.}

\item{nrow}{The number of graphs to display in a column. See 
\code{\link[ggplot2]{facet_wrap}}.}

\item{ncol}{The number of graphs to display in a row. See 
\code{\link[ggplot2]{facet_wrap}}.}
}
\description{
Create histograms or boxplots of each trait in the field book table
}
\examples{
data("fbt_sample")

# For boxplots
fb_visualize(fbt = fbt_sample)

# For histograms
fb_visualize(fbt = fbt_sample, graph = "histogram")


}
