% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fbt.R
\docType{data}
\name{fbt_sample}
\alias{fbt_sample}
\title{Field Book Table sample}
\format{A data frame with 200 rows and 14 variables}
\source{
This sample is provided in the Field Book Android app.
}
\usage{
fbt_sample
}
\description{
A \code{data.frame} to be used as a sample for demonstrations of the 
\code{fbutils} functions. It contains required columns like the unique id,
row, column, and line name, along with others such as tray number
and pedigree. The data frame also contains data for traits of different
classes (i.e. continuous, categorical, multi-categorical). Variables are as follows:
}
\details{
\itemize{
 \item plot_id The ID of the particular sample in a particular plot.
 \item row The row
 \item column The column
 \item line_name The name of the entry
 \item plot The plot number
 \item tray_row:pedigree Extra information unique to this field book table
 \item Continuous A simulated continuous trait (e.g. height)
 \item Discrete A simulated categorical/discrete trait (e.g. lodging degree)
 \item Multi A simulated multi-categorical trait (e.g. multiple disease ratings on a single plot)
 \item OneOutlier A simulated continuous trait with a single outlier
 \item SpatialGradient A simulated continuous trait with a strong gradient across rows
}
}
\keyword{datasets}
