% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbutils-package.R
\docType{package}
\name{fbutils}
\alias{fbutils}
\alias{fbutils-package}
\title{fbutils: Data Processing in Agriculture Field Trials}
\description{
The
}
\section{Functions}{

fb.utils contains several functions to aid in processing the raw data files 
from the Android Field Book App.
}

