% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{grid_cor}
\alias{grid_cor}
\title{Convenience functions}
\usage{
grid_cor(p_obs, rows, cols, grid.rows, grid.cols, layers)
}
\arguments{
\item{p_obs}{The phenotypic observations}

\item{rows}{Number of field rows}

\item{cols}{Number of field columns}

\item{grid.rows}{Number of grid rows}

\item{grid.cols}{Number of grid columns}

\item{layers}{Number of grid layers}
}
\description{
Convenience functions not generally to be called by the user.
}
