% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_heatmap.R
\name{fb_heatmap}
\alias{fb_heatmap}
\title{Generate a heatmap of field data}
\usage{
fb_heatmap(fbt, traits)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A character vector specifying the trait to plot. Defaults to all
numeric traits.}
}
\description{
Creates a heatmap image for each numeric trait in an fbt object .
}
\examples{

\dontrun{
data("fbt_sample")

# Plot the trait 'Continuous'
fb_heatmap(fbt = fbt_sample, traits = "Continuous")

# Plot all traits
fb_heatmap(fbt = fbt_sample)
}

}
