% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_outliers.R
\name{fb_outliers}
\alias{fb_outliers}
\title{Search for outliers in the field data}
\usage{
fb_outliers(fbt, traits, max.sd = 3)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A \code{character} of traits for which outliers should be 
discovered. Defaults to all numeric traits.}

\item{max.sd}{A \code{numeric} determining the number of standard deviations from
the mean to use when labeling outliers.}
}
\value{
A list of length \code{n}, where \code{n} is the number of traits. Each list
element is a data.frame containing the outliers.
}
\description{
Search for outliers in the field data
}
\details{
Outliers are defined as observations that are outside of a specified boundary
as determined by the sample mean \eqn{\mu} and standard deviation \code{s}. 
The lower limit of this boundary is \eqn{\mu - (\code{max.sd} * s)} and
the upper limit is \eqn{\mu + (\code{max.sd} * s)}.

The function prints a summary that includes the trait, sample mean, sample
standard deviation, number of missing data points, and the upper and lower
bounds of the outlier cutoff. Finally, the function prints the row from the 
fbt object in which the outlier was identified.
}
\examples{
data("fbt_sample")

outliers <- fb_outliers(fbt = fbt_sample, traits = "OneOutlier", max.sd = 3)

}
