% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_set_missing.R
\name{fb_set_missing}
\alias{fb_set_missing}
\title{Set trait values to NA}
\usage{
fb_set_missing(fbt, traits, unique_ids)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{A \code{character} vector of traits for which values should
be set to missing. Defaults to all traits.}

\item{unique_ids}{A \code{character} vector of unique identifiers for which
observations should be set to missing.}
}
\value{
An fbt object with the specified observations set to missing (i.e. \code{NA})
}
\description{
Sets observations to missing (NA) for specified traits
}
\examples{
data("fbt_sample")

# Detect the outlier in the 'OneOutlier' trait using 'fb_outliers()'
outliers <- fb_outliers(fbt = fbt_sample, traits = "OneOutlier")

# Set this outlier to NA
fbt1 <- fb_set_missing(fbt = fbt_sample, traits = "OneOutlier", unique_ids = "13RPN00010")
 
}
