% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_spatial_adj.R
\name{fb_spatial_adj}
\alias{fb_spatial_adj}
\title{Spatial adjustment of observations}
\usage{
fb_spatial_adj(fbt, traits, checks, grid.size = NULL, max.grid.size = NULL,
  use.rel.eff = TRUE)
}
\arguments{
\item{fbt}{A Field Book Table object.}

\item{traits}{Traits to be adjusted. Defaults to all numeric traits.}

\item{checks}{A \code{character} of field.book.table entry names designated
as checks.}

\item{grid.size}{A nested \code{list} of grid dimensions that define the 
moving average grid. The first layer of the list must be the same length
as the number of traits, and the names of this layer must be the trait names.
The second layer of the list must have the components \code{grid.rows}, 
\code{grid.cols}, and \code{grid.layers}. See \code{Details} for what these 
indicate. If \code{grid.size} is \code{NULL}, the grid size is optimized.}

\item{max.grid.size}{A nested \code{list}, similar to \code{grid.size}, but 
instead defining the maximum grid dimensions to use when optimizing the 
grid size. Here the components \code{grid.rows}, \code{grid.cols}, and 
\code{grid.layers} are integers defining the maximum rows, columns, and diagonal
layers to use, respectively. Use this option if the field dimensions are large.
This argument is ignored if \code{grid.size} is passed.}

\item{use.rel.eff}{\code{Logical} If \code{TRUE}, the relative efficiency is
used to determine whether the adjusted data for a trait replaces the raw data
in the \code{field.book.table}. If \code{FALSE}, both raw and adjusted data
is added to the \code{field.book.table}.}
}
\description{
Uses a moving grid to adjust the phenotypic observations
recorded in a Field Book table. This function wraps around 
functions provided in the \code{\link[mvngGrAd]{mvngGrAd}}
package.
}
\details{

}
\examples{
data("fbt_sample")

# Specify a grid size
grid.size <- list(SpatialGradient = list(grid.rows = 5, grid.cols = 5, grid.layers = 5))

adj_out <- fb_spatial_adj(fbt = fbt_sample, traits = "SpatialGradient", 
                          checks = c("Kharkof", "TAM 107", "Scout 66"), 
                          grid.size = grid.size)                         
                          
\dontrun{

# Run spatial adjustment without conditions
adj_out <- fb_spatial_adj(fbt = fbt_sample, traits = "SpatialGradient", 
                          checks = c("Kharkof", "TAM 107", "Scout 66"))
                          
# Observe the output

                          
# Restrict the grid size
max.grid.size <- list(SpatialGradient = list(grid.rows = 5, grid.cols = 5, grid.layers = 5))

# Re-run spatial adjustment
adj_out <- fb_spatial_adj(fbt = fbt_sample, traits = "SpatialGradient", 
                          checks = c("Kharkof", "TAM 107", "Scout 66"), 
                          max.grid.size = max.grid.size)
   
}                       

}
