% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{grid_cor}
\alias{grid_cor}
\title{Convenience functions}
\usage{
grid_cor(p_obs, rows, cols, grid.rows, grid.cols, layers)
}
\description{
Convenience functions not generally to be called by the user.
}
